requirements.txt.txt
pytesseract
pdf2image
Pillow
pandas
openpyxl
tk
